move jdbc test to database module
